file_line.vim: don't wipe out window
The old version used to remove another window even though I wanted to open the
file in a new window. The new version doesn't do this anymore and generally
does much less exotic stuff (it simply calls ':e' and ':call cursor(...)'). It
also doesn't check for the readability of the file anymore - it simply leaves
handling of this corner case to vim.